requirements.txt
aiographql-client==1.0.3
aiohttp==3.8.3
aiosignal==1.2.0
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot12/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
branca==0.5.0
cafeteria==0.20.0
cafeteria-asyncio==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
dadata==21.10.1
folium==0.13.0
frozenlist==1.3.1
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot12/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
graphql-core==3.2.3
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
Jinja2==3.1.2
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot12/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.23.4
pandas==1.5.1
python-dateutil==2.8.2
python-graphql-client==0.4.3
pytz==2022.5
PyYAML==5.4.1
requests==2.28.1
rfc3986==1.5.0
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot12/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.0
tqdm==4.64.1
ujson==5.5.0
urllib3==1.26.12
websockets==10.4
yarl==1.8.1

setuptools~=57.0.0